[node-hdb] Add fullVersionString and cloudVersionString to connectOptions
