+++ lib/protocol/common/ConnectOption.js
@@ -31,10 +31,12 @@ module.exports = {
   IGNORE_UNKNOWN_PARTS: 21,
   DATA_FORMAT_VERSION2: 23,
   OS_USER: 32,
+  FULL_VERSION_STRING: 44,
   REDIRECTION_TYPE: 57,
   REDIRECTED_HOST: 58,
   REDIRECTED_PORT: 59,
   ENDPOINT_HOST: 60,
   ENDPOINT_PORT: 61,
-  ENDPOINT_LIST: 62
+  ENDPOINT_LIST: 62,
+  CLOUD_VERSION_STRING: 65,
 };
