+++ lib/protocol/common/ConnectOptionType.js
@@ -34,9 +34,11 @@ ConnectOptionType[ConnectOption.ROW_AND_COLUMN_OPTIMIZED_FORMAT] = TypeCode.BOOL
 ConnectOptionType[ConnectOption.IGNORE_UNKNOWN_PARTS] = TypeCode.BOOLEAN;
 ConnectOptionType[ConnectOption.DATA_FORMAT_VERSION2] = TypeCode.INT;
 ConnectOptionType[ConnectOption.OS_USER] = TypeCode.STRING;
+ConnectOptionType[ConnectOption.FULL_VERSION_STRING] = TypeCode.STRING;
 ConnectOptionType[ConnectOption.REDIRECTION_TYPE] = TypeCode.INT;
 ConnectOptionType[ConnectOption.REDIRECTED_HOST] = TypeCode.STRING;
 ConnectOptionType[ConnectOption.REDIRECTED_PORT] = TypeCode.INT;
 ConnectOptionType[ConnectOption.ENDPOINT_HOST] = TypeCode.STRING;
 ConnectOptionType[ConnectOption.ENDPOINT_PORT] = TypeCode.INT;
 ConnectOptionType[ConnectOption.ENDPOINT_LIST] = TypeCode.STRING;
+ConnectOptionType[ConnectOption.CLOUD_VERSION_STRING] = TypeCode.STRING;
