+++ lib/protocol/part/ConnectOptions.js
@@ -104,6 +104,13 @@ function ConnectOptions() {
   // of the communication protocol instead of raising a fatal error.
   this.ignoreUnknownParts = undefined;
 
+  // This field is sent by the server and is the server version string
+  this.fullVersionString = undefined;
+
+  // This field is sent by the server and is the Cloud version that the
+  // server was released under (undefined for on-premise servers)
+  this.cloudVersionString = undefined;
+
   // The client indicates this set of understood type codes and field formats.
   // The server then defines the value according to its own capabilities,
   // and sends it back. The following values are supported:
@@ -132,10 +139,12 @@ ConnectOptions.prototype.KEYS = [
   common.ConnectOption.DATA_FORMAT_VERSION,
   common.ConnectOption.DATA_FORMAT_VERSION2,
   common.ConnectOption.OS_USER,
+  common.ConnectOption.FULL_VERSION_STRING,
   common.ConnectOption.REDIRECTION_TYPE,
   common.ConnectOption.REDIRECTED_HOST,
   common.ConnectOption.REDIRECTED_PORT,
   common.ConnectOption.ENDPOINT_HOST,
   common.ConnectOption.ENDPOINT_PORT,
-  common.ConnectOption.ENDPOINT_LIST
+  common.ConnectOption.ENDPOINT_LIST,
+  common.ConnectOption.CLOUD_VERSION_STRING,
 ];
